# App/swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "User API",
      "description": "User API Documentation"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "tags": [
      {
        "name": "users",
        "description": "User management"
      }
    ],
    "schemes": [
      "http"
    ],
    "paths": {
      "/users": {
        "post": {
          "tags": [
            "users"
          ],
          "summary": "Create a new user",
          "parameters": [
            {
              "name": "body",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/User"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "User created"
            },
            "400": {
              "description": "Invalid input"
            }
          }
        },
        "get": {
          "tags": [
            "users"
          ],
          "summary": "Get all users",
          "responses": {
            "200": {
              "description": "Successful operation"
            }
          }
        }
      },
      "/users/{id}": {
        "get": {
          "tags": [
            "users"
          ],
          "summary": "Get a user by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "type": "integer"
            }
          ],
          "responses": {
            "200": {
              "description": "Successful operation"
            },
            "404": {
              "description": "User not found"
            }
          }
        },
        "put": {
          "tags": [
            "users"
          ],
          "summary": "Update a user",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "type": "integer"
            },
            {
              "name": "body",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/User"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful operation"
            },
            "400": {
              "description": "Invalid input"
            },
            "404": {
              "description": "User not found"
            }
          }
        },
        "delete": {
          "tags": [
            "users"
          ],
          "summary": "Delete a user",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "type": "integer"
            }
          ],
          "responses": {
            "204": {
              "description": "Successful operation"
            },
            "404": {
              "description": "User not found"
            }
          }
        }
      }
    },
    "definitions": {
      "User": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      }
    }
  }
  